Rename project directory to phable
This will make it possible to install `phable-cli` but also use is as a library,
with
```python
from phable.phabricator import PhabricatorClient
```
